// web/src/shared/api/generated/model/subjectProgressResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { UnitProgressResult } from './unitProgressResult';

export interface SubjectProgressResult {
  subjectId: string;
  name: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  servableCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  masteredCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  seenCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  masteryPercent: number | string;
  units: UnitProgressResult[];
}